Restructure experimental host packages.
Move experimental bits under labs.
