+++ java/arcs/android/demo/BUILD
@@ -21,8 +21,8 @@ arcs_kt_android_library(
     resource_files = glob(["res/**/*"]),
     deps = [
         ":manifest",
-        "//java/arcs/android/host",
-        "//java/arcs/android/sdk/host",
+        "//java/arcs/android/host/parcelables",
+        "//java/arcs/android/labs/host",
         "//java/arcs/android/storage/database",
         "//java/arcs/android/util",
         "//java/arcs/core/allocator",
@@ -39,6 +39,7 @@ arcs_kt_android_library(
         "//java/arcs/core/type",
         "//java/arcs/core/util",
         "//java/arcs/jvm/util",
+        "//java/arcs/sdk/android/labs/host",
         "//java/arcs/sdk/android/storage",
         "//java/arcs/sdk/android/storage/service",
         "//java/arcs/sdk/storage",
