+++ java/arcs/android/labs/host/prod/AndroidManifest.xml
@@ -1,12 +1,12 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="arcs.android.host.prod">
+          package="arcs.android.labs.host.prod">
 
   <uses-sdk android:minSdkVersion="28" android:targetSdkVersion="28"/>
 
   <application>
     <service android:name=".ProdArcHostService" android:exported="false">
       <intent-filter>
-        <action android:name="arcs.android.host.ARC_HOST"/>
+        <action android:name="arcs.android.labs.host.ARC_HOST"/>
       </intent-filter>
     </service>
   </application>
