+++ java/arcs/android/labs/host/prod/ProdArcHostService.kt
@@ -8,13 +8,13 @@
  * grant found at
  * http://polymer.github.io/PATENTS.txt
  */
-package arcs.android.host.prod
+package arcs.android.labs.host.prod
 
 import android.content.Context
 import androidx.annotation.VisibleForTesting
 import androidx.lifecycle.Lifecycle
-import arcs.android.sdk.host.AndroidHost
-import arcs.android.sdk.host.ArcHostService
+import arcs.sdk.android.labs.host.AndroidHost
+import arcs.sdk.android.labs.host.ArcHostService
 import arcs.core.host.ArcHost
 import arcs.core.host.ParticleRegistration
 import arcs.core.host.ProdHost
