+++ java/arcs/sdk/android/labs/host/AndroidResurrectableHost.kt
@@ -8,7 +8,7 @@
  * grant found at
  * http://polymer.github.io/PATENTS.txt
  */
-package arcs.android.sdk.host
+package arcs.sdk.android.labs.host
 
 import android.content.Context
 import androidx.lifecycle.Lifecycle
