+++ javatests/arcs/android/e2e/testapp/BUILD
@@ -30,8 +30,7 @@ arcs_kt_android_library(
         ":person_arcs",
         "//java/arcs/android/devtools",
         "//java/arcs/android/devtools:aidl",
-        "//java/arcs/android/host",
-        "//java/arcs/android/sdk/host",
+        "//java/arcs/android/labs/host",
         "//java/arcs/android/storage/database",
         "//java/arcs/android/util",
         "//java/arcs/core/allocator",
@@ -48,6 +47,7 @@ arcs_kt_android_library(
         "//java/arcs/core/type",
         "//java/arcs/core/util",
         "//java/arcs/jvm/util",
+        "//java/arcs/sdk/android/labs/host",
         "//java/arcs/sdk/android/storage",
         "//java/arcs/sdk/android/storage/service",
         "//java/arcs/sdk/storage",
