+++ javatests/arcs/android/host/ArcHostHelperTest.kt
@@ -18,14 +18,14 @@ import android.os.Handler
 import android.os.ResultReceiver
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
-import arcs.android.sdk.host.ArcHostHelper
-import arcs.android.sdk.host.createGetRegisteredParticlesIntent
-import arcs.android.sdk.host.createLookupArcStatusIntent
-import arcs.android.sdk.host.createPauseArcHostIntent
-import arcs.android.sdk.host.createStartArcHostIntent
-import arcs.android.sdk.host.createStopArcHostIntent
-import arcs.android.sdk.host.createUnpauseArcHostIntent
-import arcs.android.sdk.host.toComponentName
+import arcs.sdk.android.labs.host.ArcHostHelper
+import arcs.sdk.android.labs.host.createGetRegisteredParticlesIntent
+import arcs.sdk.android.labs.host.createLookupArcStatusIntent
+import arcs.sdk.android.labs.host.createPauseArcHostIntent
+import arcs.sdk.android.labs.host.createStartArcHostIntent
+import arcs.sdk.android.labs.host.createStopArcHostIntent
+import arcs.sdk.android.labs.host.createUnpauseArcHostIntent
+import arcs.sdk.android.labs.host.toComponentName
 import arcs.core.common.ArcId
 import arcs.core.data.EntityType
 import arcs.core.data.FieldType.Companion.Text
