+++ javatests/arcs/android/host/BUILD
@@ -18,9 +18,8 @@ arcs_kt_android_test_suite(
     deps = [
         ":schemas",
         ":services",
-        "//java/arcs/android/host",
-        "//java/arcs/android/host/prod",
-        "//java/arcs/android/sdk/host",
+        "//java/arcs/android/labs/host",
+        "//java/arcs/android/labs/host/prod",
         "//java/arcs/android/storage/database",
         "//java/arcs/core/allocator",
         "//java/arcs/core/common",
@@ -39,6 +38,7 @@ arcs_kt_android_test_suite(
         "//java/arcs/core/util/testutil",
         "//java/arcs/jvm/host",
         "//java/arcs/jvm/util/testutil",
+        "//java/arcs/sdk/android/labs/host",
         "//java/arcs/sdk/android/storage",
         "//java/arcs/sdk/android/storage/service",
         "//java/arcs/sdk/android/storage/service/testutil",
@@ -79,14 +79,14 @@ arcs_kt_android_library(
     ),
     manifest = "AndroidManifest.xml",
     deps = [
-        "//java/arcs/android/host",
-        "//java/arcs/android/host/prod",
-        "//java/arcs/android/sdk/host",
+        "//java/arcs/android/labs/host",
+        "//java/arcs/android/labs/host/prod",
         "//java/arcs/core/data",
         "//java/arcs/core/host",
         "//java/arcs/core/storage",
         "//java/arcs/jvm/host",
         "//java/arcs/sdk",
+        "//java/arcs/sdk/android/labs/host",
         "//java/arcs/sdk/android/storage",
         "//java/arcs/sdk/android/storage/service/testutil",
         "//javatests/arcs/core/allocator:allocator-test-util",
