+++ javatests/arcs/android/host/parcelables/BUILD
@@ -12,7 +12,6 @@ arcs_kt_android_test_suite(
     manifest = "//java/arcs/android/common:AndroidManifest.xml",
     package = "arcs.android.host.parcelables",
     deps = [
-        "//java/arcs/android/host",
         "//java/arcs/android/host/parcelables",
         "//java/arcs/core/common",
         "//java/arcs/core/data",
